requirements.txt
streamlit>=1.22
pandas>=2.0
numpy>=1.25
pyarrow>=11.0
openpyxl>=3.1
xlrd>=2.0
plotly>=5.16
google-genai>=0.9.0
streamlit-ace>=0.5.3
python-dotenv>=1.0.0
